wipndip on contract transactions with args/kwargs. need to fix all tests pertaining to this, and change API calls in seneca to use func name args/kwargs

+++ cilantro/messages/capnp/transaction.capnp
@@ -18,7 +18,7 @@ struct ContractTransaction {
         sender @0 :Data;
         contractName @1 :Text;
         functionName @2 :Text;
-        gasSupplied @3 :UInt32;
+        gasSupplied @3 :UInt64;
         kwargs @4 :V.Map(Text, V.Value);
     }
 }
